Dynamic project columns
Refactor styles and tweak design to accommodate a flexible amount of
project columns.

We were previously using the CSS Grid module to layout the page, which
allowed us to keep the header and project items aligned on both the
X and Y axis as the page scaled. This matched the mock-up that Alex
provided, however the approach was limited in being able to dynamically
accommodate new project rows. There may be a way of doing this with
Grid, but I have dug deep enough into that spec to accomplish it that
way, if it does exist.

However, the CSS Flex module is more than capable of accommodating new
rows in the way we want, along with some slight trickery with the help
of nth-child pseudo selectors. This method allows us to have different
flexible columns depending on the amount of projects. Here's how it's
currently configured.

| Number of projects | Max number of columns |
| :----------------- | --------------------: |
| 1                  | Stretches full width  |
| 2                  | 1 row 50/50 columns   |
| 3                  | 1st row 2 up, 2nd row 100 |
| 4                  | 2 rows 50/50 columns  |
| 5                  | 1st row 33/33/33, 2nd row 50/50 |
| 6                  | both row 33/33/33 |

And so on.

Note that there is some manual process to this trickery, and will only
work elegantly like this up to 14 items. This can be amended, but the
CSS styles will need to updated with hard coded values beyond 14.

While we gained some helpful abilities with Flex, we did lose some of
Grid's super powers. The items are no longer aligned across both
dimensions across the page. This made the existing styles, which made
heavy use of large blocks of background color, looks miss-aligned and
strange (see screen shots).

Instead, I simplified the design a bit, relying on whitespace to make
each item distinct, and moving each project color to the icon, to
maintain it's personality.

I also found that we were including 2 almost identical CSS files, one
for the home page, and one for the teams page. Instead of duplicating
these, I re-factored them into one file with a common name, and included
it on both pages. This will ease maintenance.

Finally, I added the wait times project.