Add three column layout
Use `flex-basis` (via shorthand) that enables three column layout at
large widths, collapsing to two columns when there's less space, and a
single column when on mobile.

Remove the project padding from very large screens so the content will
go a little farther when we have the space.

Remove a layer of HTML nesting, and combine the inner `div` class with
the outer `section`. The styles remain the same here.

Remove the bottom border from the yellow background header on small
screens with the subtitle wraps below it to show the two are still
connected.